Move setup method to shared file
